refactor: Move rpc calls to new rpc module and simplify wrapper function
